requirements.txt
py2neo==2021.2.4
mysql-connector-python==8.0.27
pandas>=1.4.0
transformers>=4.30.0
torch>=2.0.0
sentencepiece>=0.1.99
protobuf>=3.20.0 